requirements/local.txt
# Test dependencies go here.
-r base.txt
coverage==3.7.1
django-debug-toolbar==1.0.1
django-extensions==1.3.3
model-mommy==1.2
rebar==0.1
django-shortcuts==1.5
mock==1.0.1
djet==0.0.9
